fix(readability): resolve circular import and fix global regex in punctuation spacing calc
